Create reusable submitter_information solutions and incorporate into CS and OCR.
